+++ src/Infrastructure/index.ts
@@ -1,1 +1,2 @@
 export * from './S3';
+export * from './Redis';